Move everything to backend, except one failing test
