+++ frontend/src/pages/DeviceSelector.jsx
@@ -3,6 +3,8 @@ import { useNavigate } from 'react-router-dom'
 import Navbar from "../components/Navbar"
 import logo from '../logo.png'
 import { toast } from "react-toastify";
+import { backend } from '../../wailsjs/go/models'
+import { DetectKobos, SelectKobo, PromptForLocalDBPath } from '../../wailsjs/go/backend/Backend'
 
 export default function DeviceSelector() {
   const navigate = useNavigate()
@@ -17,7 +19,7 @@ export default function DeviceSelector() {
   useEffect(() => detectDevices(), [devices.length])
 
   function detectDevices() {
-    window.go.main.KoboService.DetectKobos()
+    DetectKobos()
       .then(devices => {
         console.log(devices)
         if (devices == null) {
@@ -32,7 +34,7 @@ export default function DeviceSelector() {
   }
 
   function selectDevice(path) {
-    window.go.main.KoboService.SelectKobo(path)
+    SelectKobo(path)
       .then(error => {
         if (error === null) {
           navigate("/overview")
@@ -45,7 +47,7 @@ export default function DeviceSelector() {
   }
 
   function selectLocalDatabase() {
-    window.go.main.KoboService.PromptForLocalDBPath()
+    PromptForLocalDBPath()
       .then(error => {
         if (error === null) {
           navigate("/overview")
