--- pkg/readwise/util.go
@@ -1,9 +1,0 @@
-package readwise
-
-import "strings"
-
-func NormaliseText(s string) string {
-	s = strings.TrimSpace(s)
-	s = strings.ReplaceAll(s, "\n", " ")
-	return s
-}
